wip adding data to icons section, breaking into its own component
